requirements.txt
deepface==0.0.81
Pillow==10.2.0
ultralytics==8.1.3
omegaconf==2.3.0
streamlit==1.30.0
tf-keras
tensorflow==2.12
